test: split PDF aligned display equation coverage
